Enable Stylecop analysis (#1037)
Additionally, the following changes were made:
- analyzers are now disabled for the `Debug` and `Release` configurations;
- some StyleCop rules are suppressed for test projects;
- added a hint to switch IDE to `Verify` configuration for easier analysis fixing.

+++ Src/AutoFixtureUnitTest/DataAnnotations/RangedRequestTest.cs
@@ -17,7 +17,6 @@ namespace AutoFixtureUnitTest.DataAnnotations
             // Act & assert
             Assert.Throws<ArgumentNullException>(() =>
                 new RangedRequest(memberType, operandType, min, max));
-
         }
 
         [Fact]
@@ -37,7 +36,6 @@ namespace AutoFixtureUnitTest.DataAnnotations
             Assert.Equal(operandType, sut.OperandType);
             Assert.Equal(min, sut.Minimum);
             Assert.Equal(max, sut.Maximum);
-
         }
 
         [Fact]
@@ -203,12 +201,12 @@ namespace AutoFixtureUnitTest.DataAnnotations
         }
 
         [Theory]
-        [InlineData((int)42)]
-        [InlineData((uint)42)]
-        [InlineData((long)42)]
-        [InlineData((ulong)42)]
-        [InlineData((double)42.0)]
-        [InlineData((float)42.0f)]
+        [InlineData(42)]
+        [InlineData(42U)]
+        [InlineData(42L)]
+        [InlineData(42UL)]
+        [InlineData(42.0D)]
+        [InlineData(42.0F)]
         [InlineData("42")]
         public void ShouldCorrectlyConvertMinimum(object bounaryValue)
         {
@@ -223,12 +221,12 @@ namespace AutoFixtureUnitTest.DataAnnotations
         }
 
         [Theory]
-        [InlineData((int)42)]
-        [InlineData((uint)42)]
-        [InlineData((long)42)]
-        [InlineData((ulong)42)]
-        [InlineData((double)42.0)]
-        [InlineData((float)42.0f)]
+        [InlineData(42)]
+        [InlineData(42U)]
+        [InlineData(42L)]
+        [InlineData(42UL)]
+        [InlineData(42.0D)]
+        [InlineData(42.0F)]
         [InlineData("42")]
         public void ShouldCorrectlyConvertMaximum(object bounaryValue)
         {
@@ -300,7 +298,6 @@ namespace AutoFixtureUnitTest.DataAnnotations
             Assert.Contains("Int64", stringResult);
             Assert.Contains("42", stringResult);
             Assert.Contains("100", stringResult);
-
         }
     }
 }
